+++ starkinfra/__init__.py
@@ -1,4 +1,4 @@
-version = "0.3.1"
+version = "0.4.0"
 language = "en-US"
 timeout = 15
 user = None